+++ modules/task_2/purtova_k_shell_sort_batcher_merge/shell_sort.h
@@ -4,7 +4,7 @@
 
 #include <vector>
 
-void Shell(const std::vector<int>& A, int n);
+std::vector<int> Shell(const std::vector<int>& A, int n);
 std::vector <int> MergerEvenOdd(const std::vector<int>& arr1, const std::vector<int>& arr2, int i1, int i2, int size_result);
 std::vector<int> BatcherEvenNumber(const std::vector<int>& arr1, const std::vector<int>& arr2);
 std::vector<int> BatcherOddNumber(const std::vector<int>& arr1, const std::vector<int>& arr2);
@@ -13,4 +13,4 @@ std::vector<int> ShellWithBatcher(const std::vector<int>& arr, int size, int num
 std::vector<int> ShellWithBatcherOpenMP(const std::vector<int>& arr, int size, int number);
 std::vector<int> VectorRandomizer(int size);
 std::vector <std::vector<int>> splitVector(const std::vector<int>& vec, size_t number);
-#endif  // MODULES_TASK_1_PURTOVA_K_SHELL_SORT_BATCHER_MERGE_SHELL_SORT_H_
+#endif  // MODULES_TASK_2_PURTOVA_K_SHELL_SORT_BATCHER_MERGE_SHELL_SORT_H_
